Cleanup. Remove not used code
